Move DB tasks to model/dao object
